test(react-mcp): share rejection capture helper
