+++ packages/react-mcp/src/primitives/server/test-utils.ts
@@ -1,0 +1,12 @@
+export const captureUnhandledRejections = async (callback: () => void) => {
+  const reasons: unknown[] = [];
+  const listener = (reason: unknown) => reasons.push(reason);
+  process.on("unhandledRejection", listener);
+  try {
+    callback();
+    await new Promise((resolve) => setTimeout(resolve, 0));
+    return reasons;
+  } finally {
+    process.off("unhandledRejection", listener);
+  }
+};
